+++ CHANGELOG.rst
@@ -13,11 +13,17 @@ Change Log
 Unreleased
 ~~~~~~~~~~
 
+[5.0.0] - 2021-03-18
+~~~~~~~~~~~~~~~~~~~~
+
+* BREAKING CHANGE: Add linter for invalid imports from Django Waffle (`import waffle` and `from waffle import ...`). Instead, developers should import toggle objects from `edx_toggles.toggles`.
+* BREAKING CHANGE: Add linter for missing feature toggle annotations ("toggle-missing-annotation"). Check `this howto <https://edx.readthedocs.io/projects/edx-toggles/en/latest/how_to/documenting_new_feature_toggles.html>`__ for more information on writing toggle annotations.
+* Fix duplicate annotation errors.
+
 [4.1.1] - 2021-03-16
 ~~~~~~~~~~~~~~~~~~~~
 
 * Fixed lint amnesty breakage on line continuation
-
 
 [4.1.0] - 2021-02-24
 ~~~~~~~~~~~~~~~~~~~~